Add missing pow function implementation. Missed on merge from mingw. Additionally cast arguments of pow in powf to double.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@426 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-crt/math/powf.c
@@ -5,4 +5,6 @@
  */
 #include <math.h>
 float powf (float x, float y)
-  {return (float) pow (x, y);}
+{
+  return (float) pow ((double) x, (double) y);
+}
